requirements.txt
Flask==2.2.5
flask-pymongo==2.3.0
python-dotenv==1.0.0
dnspython==2.2.1